Move azul.service.hca_response_v5 to into plugin
