+++ test/service/test_response.py
@@ -51,7 +51,7 @@ from azul.indexer.index_service import (
 from azul.logging import (
     configure_test_logging,
 )
-from azul.service.hca_response_v5 import (
+from azul.plugins.metadata.hca.stages.response import (
     Pagination,
     SearchResponseFactory,
 )
